% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprClass.R
\name{as.spr}
\alias{as.spr}
\title{Convert data.frame to spr}
\usage{
as.spr(data, axis, units, sample, history = "")
}
\arguments{
\item{data}{data.frame with 2 columns (spectralAxis and response)}

\item{axis}{character vector containing axis type (spectralAxis may have "wavenumber", "wavelength" or "frequency" type, while response may have "absorbance", "transmittance" or "SB" type)}

\item{units}{character vector containing axis units (WIP)}

\item{sample}{name of sample}

\item{history}{character vector containing changes to spectra. When creating a new spr object it is recommended to keep it empty (default behaviour)}
}
\description{
Function to convert data.frame to spr, adding necessary information about the spectra.
}
