% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGfilter.R
\name{SG.filter}
\alias{SG.filter}
\title{Savitzki-Golay filter for FTIR}
\usage{
SG.filter(spr, N, M, H, padding = TRUE)
}
\arguments{
\item{spr}{object of class spr}

\item{N}{polynomial order}

\item{M}{window half-width}

\item{H}{pseudo-inverse of the Valdemondo matrix. Overrides given values of N,M.}

\item{padding}{if TRUE, add a padding of length M}
}
\description{
Savitzki-Golay filter for FTIR
}
