% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprClass.R
\name{show,spectra-method}
\alias{show,spectra-method}
\title{Present important spr information when printing}
\usage{
\S4method{show}{spectra}(object)
}
\arguments{
\item{object}{spr object}
}
\description{
Present important spr information when printing
}
